add priorities to testing options

+++ localtest.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 set -e
+#set -x
 
 dir=$(dirname "$0")
 
@@ -8,16 +9,18 @@ sudo rm -f $dir/coverage/* || true
 
 docker images alpine | grep alpine > /dev/null || docker pull alpine
 docker build -t droutertest -f $dir/Dockertest $dir
-echo "$@"
 
-# test pid1
-docker run -it --name=drntest_drouter --privileged --rm -e TERM=xterm -e NO_TEST_SETUP=1 -e TEST_NO_HOST_PID=1 -v /var/run/docker.sock:/var/run/docker.sock -v $(pwd)/$dir/coverage:/coverage droutertest "go test github.com/TrilliumIT/docker-drouter/drouter -timeout 20m -coverprofile=/coverage/pid1.out -run=TestPid1"
+if [[ -z "$@" ]]; then
+	# test pid1
+	docker run -it --name=drntest_drouter --privileged --rm -e TERM=xterm -e NO_TEST_SETUP=1 -e TEST_NO_HOST_PID=1 -v /var/run/docker.sock:/var/run/docker.sock -v $(pwd)/$dir/coverage:/coverage droutertest "go test github.com/TrilliumIT/docker-drouter/drouter -timeout 20m -coverprofile=/coverage/pid1.out -run=TestPid1"
 
-# test no docker socket
-docker run -it --name=drntest_drouter --privileged --rm --pid=host -e TERM=xterm -e NO_TEST_SETUP=1 -e TEST_NO_SOCKET=1 -v $(pwd)/$dir/coverage:/coverage droutertest "go test github.com/TrilliumIT/docker-drouter/drouter -timeout 20m -coverprofile=/coverage/nosocket.out -run=TestNoSocket"
+	# test no docker socket
+	docker run -it --name=drntest_drouter --privileged --rm --pid=host -e TERM=xterm -e NO_TEST_SETUP=1 -e TEST_NO_SOCKET=1 -v $(pwd)/$dir/coverage:/coverage droutertest "go test github.com/TrilliumIT/docker-drouter/drouter -timeout 20m -coverprofile=/coverage/nosocket.out -run=TestNoSocket"
+fi
 
+TEST_CMD="go test github.com/TrilliumIT/docker-drouter/drouter -timeout 20m -coverprofile=/coverage/main.out $@"
 # test main
-docker run -it --name=drntest_drouter --privileged --rm -e TERM=xterm --pid=host -v /var/run/docker.sock:/var/run/docker.sock -v $(pwd)/$dir/coverage:/coverage droutertest "go test github.com/TrilliumIT/docker-drouter/drouter -timeout 20m -coverprofile=/coverage/main.out $@"
+docker run -it --name=drntest_drouter --privileged --rm -e TERM=xterm --pid=host -v /var/run/docker.sock:/var/run/docker.sock -v $(pwd)/$dir/coverage:/coverage droutertest "$TEST_CMD"
 
 # remove dangling images
 [ -z $(docker images -q -f dangling=true | head -1) ] || docker rmi $(docker images -q -f dangling=true) > /dev/null
